+++ rororo/__init__.py
@@ -10,5 +10,5 @@ applications. Supports `aiohttp.web <http://aiohttp.readthedocs.org/>`_,
 """
 
 __author__ = 'Igor Davydenko'
-__license__ = 'BSD'
+__license__ = 'BSD-3-Clause'
 __version__ = '1.2.0a0'